FEAT: 기본 라우팅 설정 및 컴포넌트 쪼개기
