+++ src/components/common/Loading.tsx
@@ -1,0 +1,4 @@
+const Loading = () => {
+    return <><div>loading.... </div></>
+}
+export default Loading;
