+++ src/components/layout/AppLayout.tsx
@@ -1,0 +1,6 @@
+
+const AppLayout = () => {
+    return <></>
+}
+
+export default AppLayout;
