+++ src/components/layout/index.ts
@@ -1,0 +1,1 @@
+export {default as AppLayout} from './AppLayout'
