+++ src/pages/NotFound.tsx
@@ -1,0 +1,5 @@
+const NotFound = () => {
+  return <div>404 not found 페이지를 찾을 수 없습니다.</div>;
+};
+
+export default NotFound;
