+++ src/pages/index.ts
@@ -1,0 +1,2 @@
+export { default as OrderPage } from './Order'
+export { default as NotFoundPage } from './NotFound'
